# BUG/PERF: MaskedArray.searchsorted(np.nan)