--- a/pandas/tests/arrays/masked_shared.py
+++ b/pandas/tests/arrays/masked_shared.py
@@ -55,6 +55,14 @@ class ComparisonOps(BaseOpsUtil):
 class NumericOps:
     # Shared by IntegerArray and FloatingArray, not BooleanArray
 
+    def test_searchsorted_nan(self, dtype):
+        # The base class casts to object dtype, for which searchsorted returns
+        #  0 from the left and 10 from the right.
+        arr = pd.array(range(10), dtype=dtype)
+
+        assert arr.searchsorted(np.nan, side="left") == 10
+        assert arr.searchsorted(np.nan, side="right") == 10
+
     def test_no_shared_mask(self, data):
         result = data + 1
         assert not tm.shares_memory(result, data)