@jreback 2022-01-08
>  # The base class casts to object dtype, for which searchsorted returns
        #  0 from the left and 10 from the right.

i think needs a whatsnew? as this is the bug right